update auth function organization
