+++ src/auth/index.js
@@ -1,4 +1,20 @@
 'use strict'
 
-export * from './authMessages'
-export * from './authBrowser'
+export {
+  isUserSignedIn, redirectUserToSignIn, getAuthResponseToken, isSignInPending,
+  signUserIn, loadUserData, signUserOut
+} from './authApp'
+export {
+  makeAuthRequest, makeAuthResponse
+} from './authMessages'
+export {
+  getAuthRequestFromURL, fetchAppManifest, redirectUserToApp
+} from './authProvider'
+export {
+  makeCoreSessionRequest, sendCoreSessionRequest, getCoreSession
+} from './authSession'
+export {
+  verifyAuthRequest, verifyAuthResponse,
+  isExpirationDateValid, isIssuanceDateValid, doPublicKeysMatchUsername,
+  doPublicKeysMatchIssuer, doSignaturesMatchPublicKeys
+} from './authVerification'
